requirements.txt
keyboard==0.13.5
colorama==0.4.6
PyYAML==6.0.1
